refactor(project): reorganize file structure into pages and sections
- Rename `src/pages` to `src/sections` for reusable UI sections (Hero, About, Tools, Contact, etc.)
- Create new `src/pages` directory containing `PortfolioPage.tsx` and `CVPage.tsx` as route-level components
