+++ src/App.tsx
@@ -2,10 +2,10 @@ import { useEffect } from "react";
 
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
-import Portfolio from "./components/Portfolio";
+import Portfolio from "./pages/Portfolio";
 
 import Lenis from "lenis";
-import CVPage from "./pages/CvPage";
+import CVPage from "./sections/CvPage";
 
 const App = () => {
   // Lenis Logic
